refactor: deduplicate namespace flags and regenerate README via make readme
Maintainability and readability improvements from the deep review:

- Introduce a shared namespaces struct holding the target and storage
  namespace, with the storage() fallback method and a single
  addNamespaceFlags helper registering -n/--namespace and
  --storage-namespace (including flag help texts and env var defaults).
  Embed it into diffCmd, revision and rollback, removing the three
  copies of the getStorageNamespace method and the three copies of the
  flag registration boilerplate
- Remove the redundant diffCmd namespace initialization in
  newChartCommand: the flag default already assigns the same
  HELM_NAMESPACE value at registration time
- Add scripts/gen-readme.sh plus make readme / make verify-readme to
  regenerate the cobra flag tables in README.md from the actual --help
  output, and verify them in CI, so the documented flags can no longer
  drift (the release table had drifted: missing --kube-context and
  --show-secrets-decoded rows, now fixed)
